Making bt result type checking more consistent with the rest of the api usage (e.g. via bt.*)
