+++ backtrader/plot/bokeh/datatable.py
@@ -1,8 +1,12 @@
 import backtrader as bt
 from collections import OrderedDict
-from bokeh.models import ColumnDataSource, Paragraph, TableColumn, DataTable, DateFormatter, NumberFormatter, StringFormatter, Widget
 from enum import Enum
 from .utils import get_params_str
+
+try:
+    from bokeh.models import ColumnDataSource, Paragraph, TableColumn, DataTable, DateFormatter, NumberFormatter, StringFormatter, Widget
+except ImportError as e:
+    raise ImportError('Bokeh seems to be missing. Needed for bokeh plotting support')
 
 
 class ColummDataType(Enum):
