+++ backtrader/plot/bokeh/utils.py
@@ -1,16 +1,24 @@
-from jinja2 import Environment, PackageLoader
 from datetime import datetime
-import matplotlib.colors
 import backtrader as bt
 import math
 import numbers
 import logging
 
 try:
+    from jinja2 import Environment, PackageLoader
+except ImportError:
+    raise ImportError('jinja2 seems to be missing. Needed for bokeh plotting support')
+
+try:
     import pandas
 except ImportError:
-    raise ImportError(
-        'Pandas seems to be missing. Needed for bokeh plotting support')
+    raise ImportError('Pandas seems to be missing. Needed for bokeh plotting support')
+
+try:
+    import matplotlib.colors
+except ImportError:
+    raise ImportError('matplotlib seems to be missing. Needed for bokeh plotting support')
+
 
 _logger = logging.getLogger(__name__)
